+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Remove ticket number from PR title when there is no ticket key
 - Add `PR_TICKET_PREFIX_TEXT`
 - Refactor unique src file into multiple (single scope/responsibility) files
+- Add option `-t|--title` to generate a branch name based on the PR title.
 
 ## 0.4 - 2024-09-06